tests/wpt/web-platform-tests/typedarrays/ArrayBuffer_constructor.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Typed Arrays Test: ArrayBuffer constructor</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.khronos.org/registry/typedarray/specs/latest/#5">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

var args = [
  /* numbers */
  [NaN, 0], [+0, 0], [-0, 0], [1, 1],
  [-0.4, 0], [-0.9, 0], [1.1, 1], [2.9, 2],
  /* strings */
  ["1", 1], ["1e2", 100],
  /* null, undefined, booleans */
  [undefined, 0], [null, 0], [false, 0], [true, 1]
];

args.forEach(function (arg, i) {
  test(function () {
    var abuffer = new ArrayBuffer(arg[0]);
    assert_equals(abuffer.byteLength, arg[1]);
  }, "The argument " + format_value(arg[0]) + " should be interpreted as " +
     arg[1] + " for ArrayBuffer constructor." + i);
});

var invalid_args = [
  +Infinity, -Infinity, 2**64, -(2**64),
  2**53, -(2**53), -(2**32), -(2**31),
  -1, -2.5,
];

invalid_args.forEach(function (arg, i) {
  test(function () {
    assert_throws(new RangeError, function() {
      var abuffer = new ArrayBuffer(arg);
    });
  }, "The argument " + format_value(arg) + " should cause a RangeError exception when passed to the ArrayBuffer constructor." + i);
});

</script>
